PR comments and move UpdateContacts details test into shared package
